application of Convolutional Neural Networks (CNNs)
This project demonstrates the application of Convolutional Neural Networks (CNNs) using Tensor Flow for flower image classification. Leveraging deep learning techniques, the model is trained to recognize and categorize different flower species from images. With practical code examples, it serves as an educational resource for those interested in computer vision, image classification, and Tensor Flow. Perfect for beginners looking to delve into the world of deep learning and image recognition or for developers wanting to expand their knowledge in this domain. Explore the beauty of AI-powered flower recognition with Tensor Flow.